added the base of auth middleware
