+++ pkg/api/helpers.go
@@ -1,7 +1,9 @@
 package api
 
 import (
+	"akshidas/e-com/pkg/utils"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"net/http"
 	"strconv"
@@ -15,6 +17,27 @@ type ApiError struct {
 
 type ApiResponse struct {
 	Data any `json:"data"`
+}
+
+func IsAuthenticated(f apiFunc) apiFunc {
+	return func(w http.ResponseWriter, r *http.Request) error {
+		authtoken := r.Header.Get("Authorization")
+		token, err := utils.ValidateJWT(authtoken)
+		if err != nil {
+			accessDenied(w)
+		}
+
+		if !token.Valid {
+			accessDenied(w)
+		}
+		return f(w, r)
+	}
+
+}
+
+func accessDenied(w http.ResponseWriter) error {
+	return writeError(w, http.StatusUnauthorized, errors.New("access denied"))
+
 }
 
 func RouteHandler(f apiFunc) http.HandlerFunc {
