+++ pkg/server/logger.go
@@ -1,0 +1,21 @@
+package server
+
+import (
+	"log"
+	"net/http"
+	"time"
+)
+
+type Logger struct {
+	handler http.Handler
+}
+
+func (l *Logger) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	start := time.Now()
+	l.handler.ServeHTTP(w, r)
+	log.Printf("%s %s %v", r.Method, r.URL.Path, time.Since(start))
+}
+
+func NewLogger(handleToWrap http.Handler) *Logger {
+	return &Logger{handleToWrap}
+}
